fix: discount percentage and amount datatype change and razorpay live key generated

+++ Backend/Agronexis.Model/ResponseModel/PriceResponseModel.cs
@@ -13,8 +13,8 @@ namespace Agronexis.Model.ResponseModel
         public string CurrencyCode { get; set; }
         public decimal Price { get; set; }
         public bool IsDiscounted { get; set; }
-        public int DiscountPercentage { get; set; }
-        public int DiscountedAmount { get; set; }
+        public decimal DiscountPercentage { get; set; }
+        public decimal DiscountedAmount { get; set; }
         public string SkuNumber { get; set; }
         public Guid WeightId { get; set; }
         public decimal? WeightValue { get; set; }
